Cleanup filters for PolyData
