+++ pyvista/core/__init__.py
@@ -1,5 +1,5 @@
 from .common import Common
 from .composite import MultiBlock
-from .filters import CompositeFilters, DataSetFilters
+from .filters import CompositeFilters, DataSetFilters, PolyDataFilters, UnstructuredGridFilters
 from .grid import Grid, RectilinearGrid, UniformGrid
 from .pointset import PointGrid, PolyData, StructuredGrid, UnstructuredGrid
